DEV: Rename _class.py to error_estimates.py
